Move the MockFS implementation to mockfs.mfs and eliminate globals
mockfs.install() now returns the newly-created mockfs instance.
The original API used module globals and always created a singleton
instance within the mockfs module.  We now create a new instance
every time.

The documentation examples now reflect the updated API.
This makes things more transparent and removes the need for us
to provide the singleton() function and the global variables
that come along with it.

Signed-off-by: David Aguilar <davvid@gmail.com>