fix various PEP8 issues
Most of the changes are harmless.

Some of them actually alter the very-old style that probably
was copied from original docstrings from stdlib.
None of the tests actually check if mocked IO has the same
docstrings as stdlib IO so I guess that's not that important

There is also one old exception syntax change but I don't think
that's a factor as python2.5 is not that important anymore

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>